Update Fruit database

+++ Models/Fruit.cs
@@ -1,10 +1,8 @@
-using System.ComponentModel.DataAnnotations;
-
 namespace AspNetCoreWebApi.Models
 {
     public class Fruit
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string No { get; set; }
         public string Description { get; set; }
     }
